[vme] use long for waitIrq timeout parameter
On x86_64 this parameter actually needs to be 64bit,
that means `long` is actually the correct type!

+++ vmelib/vmelib.h
@@ -98,7 +98,7 @@ public:
   int setupIrq(int image, unsigned int irqLevel, unsigned int statusID, unsigned int addrSt, unsigned int valSt, unsigned int addrCl, unsigned int valCl);
   int freeIrq(int image, unsigned int irqLevel, unsigned int statusID);
   int waitIrq(unsigned int irqLevel, unsigned int statusID);
-  int waitIrq(unsigned int irqLevel, unsigned int statusID, uint32_t timeout);
+  int waitIrq(unsigned int irqLevel, unsigned int statusID, long timeout);
   int generateVmeIrq(unsigned int irqLevel, unsigned int statusID);
 
   // Mailbox use
